tsuru man page file written using troff style
The number 8 in the file name refers to section number of UNIX like system
man pages.
This number means that this man page is from a System Administration Command.
In the install process, this file must be placed at system man page directory.
Normally it should be copied to '/usr/share/man/man8/'