auth: Keep .tsuru-token on user-remove
Don't delete `~/.tsuru-token` when deleting a user with `user-remove`. So
that an administrator deleting other users doesn't have to re-login after
each command.

I did think about retaining the existing behaviour for the "without args"
case where you are deleting yourself. However it ever so slightly increased
the code complexity and the test may have produced false-negatives in the
future, which didn't seem worth it when the user's token would be
invalidated and they'd be prompted to login the next time anyway.

Fixes tsuru/tsuru#1235