installer: uses instance private in some clouds
This adds support for instances private ips. When available,
the private IP will be used as address of the tsuru docker node and
the swarm advertise address, so nodes can talk to each other without
the need of aditional security configuration. The private IP interface,
can be configured using the config "driver:private-ip-interface",
the method "GetPrivateIPInterface" must contain fallbacks for the core
drivers(currently, only amazonec2 is implemented)..

While the Driver
interface does not provide a way to get the private IP address.