fix tsuru issue #1442 - token-show and user-remove commands correct passing email argument
When you execute 'tsuru token-show -u test+testing@example.com' the
email argument wasn't encoded correctly as a query string parameter to
the tsuru server.

As a result the tsuru server was getting "test testing@example.com" as
the email value on this case.

Fixed that query string encoding on "token-show" and "user-remove"
commands with this commit.